+++ src/modules/previews/interfaces/preview.ts
